Added Error Popup Message
Added a JS named popup that will control any kind of popup message we
need anywhere all you need to do is call the function popup(argument);
note the argument needs to be a string literal or a variable containing
a string.  I have already used this on the admin login JS if you need
guidance.  if you try to push submit on the admin login in will popup an
error message.  I can not however get the font to change to black for
visibility, it's currently white text on a yellow background popup box.